+++ omni-led-lib/src/events/mod.rs
@@ -1,4 +1,5 @@
 pub mod event_loop;
 pub mod event_queue;
 pub mod events;
+pub mod proto_to_lua;
 pub mod shortcuts;
